Corrected the logic for backfill.

See merge request TeamProject36/PROJECT!39

+++ frontend/src/routes/TypeViewer.js
@@ -8,31 +8,37 @@ import {
 import { useEffect, useState } from 'react';
 import { Link as RouteLink } from 'react-router-dom';
 import { deleteType, fetchAllTypes } from '../api';
+import TypeMethodManager from '../components/TypeMethodManager';
 import useAuth from '../hooks/useAuth';
 
 const TypeViewer = () => {
 	const { user } = useAuth();
 	const [toggle, set_toggle] = useBoolean();
+	const [types, set_types] = useState([]);
+	const [highest, set_highest] = useState([]);
 
 	useEffect(() => {
 		async function load_allTypes() {
 			let data = await fetchAllTypes(res => res.data);
+			set_highest(TypeMethodManager.assignHighest(data));
 			set_types(data);
 		}
 		load_allTypes();
 	}, [toggle]);
 
-	const [types, set_types] = useState([]);
-
 	const deleteThis = (type) => {
-		deleteType(type.typeId).then(data => {
-			if (data.wasAllowed == false) {
-				alert('Type ' + type.typeName + ' is depended upon, can not be deleted.');
-			}
-			else {
-				set_toggle.toggle();
-			}
-		});
+		let text = 'Waring!\nThis will delete all versions of this type';
+		let doDelete = window.confirm(text);
+		if (doDelete) {
+			deleteType(type.typeId).then(data => {
+				if (!data.wasAllowed) {
+					alert('Type ' + type.typeName + ' is depended upon, can not be deleted.');
+				}
+				else {
+					set_toggle.toggle();
+				}
+			});
+		}
 
 	};
 
@@ -40,20 +46,23 @@ const TypeViewer = () => {
 		<VStack>
 			<TableContainer>
 				<Table varient='simple'>
-					<TableCaption placement='top' color='white'>Types</TableCaption>
+					<TableCaption placement='top'>Types</TableCaption>
 					<Thead>
 						<Tr>
-							<Th color='white'>Type</Th>
-							<Th color='white'>Attributes</Th>
-							<Th color='white'>Attribute Data Type</Th>
-							<Th color='white'>Delete</Th>
+							<Th>Type</Th>
+							<Th>Version</Th>
+							<Th>Attributes</Th>
+							<Th>Attribute Data Type</Th>
+							<Th>Edit</Th>
+							<Th>Delete</Th>
 						</Tr>
 					</Thead>
 					<Tbody>
-						{types.map((types) => {
+						{types.map((types, index) => {
 							return (
-								<Tr key={types.typeName}>
+								<Tr key={types.versionId}>
 									<Td>{types.typeName}</Td>
+									<Td>{types.versionNumber}</Td>
 									<Td>
 										{types.metadata.map((metadata) => {
 											return (
@@ -73,7 +82,16 @@ const TypeViewer = () => {
 										})}
 									</Td>
 									<Td>
-										<Button onClick={() => deleteThis(types)}>Delete</Button>
+										{highest[index] === types.versionNumber &&
+											<RouteLink to={`./${types.versionId}`}>
+												<Button>Edit</Button>
+											</RouteLink>
+										}
+									</Td>
+									<Td>
+										{highest[index] === types.versionNumber &&
+											<Button onClick={() => deleteThis(types)}>Delete</Button>
+										}
 									</Td>
 								</Tr>
 							);
